+++ src/auth.js
@@ -1,13 +1,16 @@
 "use strict";
 
+//process.env['NODE_TLS_REJECT_UNAUTHORIZED'] = '0'; // Ignore 'UNABLE_TO_VERIFY_LEAF_SIGNATURE' authorization error
+
 var request = require('request').defaults({
     headers: {
-        "x-tesla-user-agent": "TeslaApp/3.10.8-421/adff2e065/android/8.1.0",
-        "user-agent": "Mozilla/5.0 (Linux; Android 8.1.0; Pixel XL Build/OPM4.171019.021.D1; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/68.0.3440.91 Mobile Safari/537.36",
-        "x-requested-with": "com.teslamotors.tesla"
+        "Accept" : "*/*"
     },
     gzip: true,
+//    timeout: 60000,
+//    followRedirect: false,
     jar: true
+//    ,proxy: "http://127.0.0.1:8888" // Note the fully-qualified path to Fiddler proxy. No "https" is required, even for https connections to outside.
 });
 var crypto = require('crypto');
 var Promise = require('promise');
@@ -16,7 +19,7 @@ exports.login = function login(credentials, callback) {
     var codeVerifier = generateCodeVerifier();
     var codeChallenge = generateCodeChallenge(codeVerifier);
     var queryString = {
-        audience: '',
+//        audience: '',
         client_id: 'ownerapi',
         code_challenge: codeChallenge,
         code_challenge_method: 'S256',
@@ -25,7 +28,8 @@ exports.login = function login(credentials, callback) {
         redirect_uri: 'https://auth.tesla.com/void/callback',
         response_type: 'code',
         scope: 'openid email offline_access',
-        state: generateCodeChallenge(generateCodeVerifier())
+        state: generateCodeChallenge(generateCodeVerifier()),
+        login_hint: credentials.identity
     };
     var transactionId = null;
     var loginHost = null;
@@ -33,7 +37,7 @@ exports.login = function login(credentials, callback) {
 
     req({
         method: 'GET',
-        url: 'https://auth-global.tesla.com/oauth2/v3/authorize',
+        url: 'https://auth.tesla.com/oauth2/v3/authorize',
         qs: queryString,
         headers: {
             "sec-fetch-site": "none",
@@ -102,6 +106,11 @@ exports.login = function login(credentials, callback) {
             url: (url.query.issuer || 'https://auth.tesla.com/oauth2/v3') + '/token',
             jar: false,
             json: true,
+            headers: {
+                "Accept": "*/*",
+                "Content-Type" : "application/json",
+                "Connection" : "keep-alive"
+            },
             body: {
                 grant_type: 'authorization_code',
                 client_id: 'ownerapi',
@@ -207,7 +216,7 @@ function bearerForAccessToken(bearerResult) {
 function generateCodeVerifier() {
     // Tesla might use something more sophisticated, but in my experience it's a 112-char alphanumeric string so let's just do that
     var chars = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890';
-    var random = crypto.randomBytes(112);
+    var random = crypto.randomBytes(86);
     var output = '';
     for (var i = 0; i < random.length; i++) {
         output += chars[random[i] % chars.length];
@@ -245,7 +254,7 @@ exports.refresh = function refresh(refresh_token, callback) {
 function req(parameters) {
     return new Promise(function (resolve, reject) {
         request(parameters, function (error, response, body) {
-            if (error || response.statusCode >= 400) {
+            if (error /*|| response.statusCode >= 400*/) {
                 return reject(error || new Error("HTTP error " + response.statusCode));
             }